samplesize warning
git-svn-id: svn://svn.icculus.org/netradiant/trunk@102 61c419a2-8eb2-4b30-bcec-8cead039b335

+++ tools/quake3/q3map2/lightmaps_ydnar.c
@@ -694,8 +694,8 @@ qboolean AddSurfaceToRawLightmap( int num, rawLightmap_t *lm )
 			info->maxs[0],
 			info->maxs[1],
 			info->maxs[2],
-			lm->sampleSize,
-			sampleSize);
+			(double) lm->sampleSize,
+			(double) sampleSize);
 	}
 	
 	/* set actual sample size */
